♻️ refactor : Task와 UseCase 분리, kafka event 폴더 이동
