+++ src/main/java/com/zipte/notifications/server/adapter/in/consumer/CommentEventConsumer.java
@@ -1,9 +1,9 @@
 package com.zipte.notifications.server.adapter.in.consumer;
 
-import com.zipte.notifications.server.application.port.in.event.AddCommentEventUseCase;
-import com.zipte.notifications.server.application.port.in.event.RemoveCommentEventUseCase;
-import com.zipte.notifications.server.domain.CommentEvent;
-import com.zipte.notifications.server.domain.CommentEventType;
+import com.zipte.notifications.server.application.port.in.task.AddCommentNotificationTask;
+import com.zipte.notifications.server.application.port.in.task.RemoveCommentNotificationTask;
+import com.zipte.notifications.server.adapter.in.consumer.dto.CommentEvent;
+import com.zipte.notifications.server.adapter.in.consumer.dto.CommentEventType;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.context.annotation.Bean;
@@ -16,8 +16,8 @@ import java.util.function.Consumer;
 @RequiredArgsConstructor
 public class CommentEventConsumer {
 
-    private final AddCommentEventUseCase addTask;
-    private final RemoveCommentEventUseCase removeTask;
+    private final AddCommentNotificationTask addTask;
+    private final RemoveCommentNotificationTask removeTask;
 
     @Bean("comment")
     public Consumer<CommentEvent> comment() {
