+++ src/main/java/com/zipte/notifications/server/adapter/in/consumer/dto/CommentEventType.java
@@ -1,0 +1,6 @@
+package com.zipte.notifications.server.adapter.in.consumer.dto;
+
+public enum CommentEventType {
+    ADD,
+    REMOVE
+}
