+++ src/main/java/com/zipte/notifications/server/adapter/in/consumer/dto/PropertyEventType.java
@@ -1,0 +1,6 @@
+package com.zipte.notifications.server.adapter.in.consumer.dto;
+
+public enum PropertyEventType {
+    ADD,
+    REMOVE
+}
