--- src/main/java/com/zipte/notifications/server/domain/CommentEventType.java
@@ -1,6 +1,0 @@
-package com.zipte.notifications.server.domain;
-
-public enum CommentEventType {
-    ADD,
-    REMOVE
-}
