--- src/main/java/com/zipte/notifications/server/domain/PropertyEventType.java
@@ -1,6 +1,0 @@
-package com.zipte.notifications.server.domain;
-
-public enum PropertyEventType {
-    ADD,
-    REMOVE
-}
